Refactor Config Platform adapter
